Manage Cron Jobs within the JS Process Fixes #3183
