+++ .changeset/lazy-pans-design.md
@@ -1,0 +1,5 @@
+---
+"frontend": minor
+---
+
+Removes the Cron-Webhook service in favour of the croner library
