+++ .env.example
@@ -15,5 +15,3 @@ GOOGLE_PRIVATE_KEY=
 GOOGLE_CALENDAR_ID=
 
 ADMIN_ROLE_ID=
-
-WEBHOOK_SERVER=
