+++ package.json
@@ -1,100 +1,99 @@
 {
-	"name": "frontend",
-	"private": false,
-	"version": "5.14.16",
-	"type": "module",
-	"scripts": {
-		"dev": "bunx vite",
-		"build": "bun vite build",
-		"start": "PORT=1420 bun run .output/server/index.mjs",
-		"check": "bun biome check ./src",
-		"typecheck": "bun tsc --noEmit",
-		"knip": "knip-bun --config knip.json"
-	},
-	"main": ".output/server/index.mjs",
-	"files": [
-		"drizzle",
-		".output"
-	],
-	"dependencies": {
-		"@discordjs/builders": "1.13.1",
-		"@discordjs/core": "2.4.0",
-		"@discordjs/rest": "2.6.0",
-		"@emotion/react": "11.14.0",
-		"@emotion/styled": "11.14.1",
-		"@fontsource/roboto": "5.2.9",
-		"@graphql-yoga/subscription": "5.0.5",
-		"@hono/zod-validator": "0.7.6",
-		"@hookform/resolvers": "5.2.2",
-		"@keyv/postgres": "2.2.2",
-		"@lucia-auth/adapter-drizzle": "1.1.0",
-		"@mui/icons-material": "7.3.6",
-		"@mui/material": "7.3.6",
-		"@mui/x-date-pickers": "8.23.0",
-		"@t3-oss/env-core": "0.13.10",
-		"@tanstack/match-sorter-utils": "8.19.4",
-		"@tanstack/react-query": "5.90.12",
-		"@tanstack/react-query-devtools": "5.91.1",
-		"@tanstack/react-router": "1.143.11",
-		"@tanstack/react-router-devtools": "1.143.11",
-		"@tanstack/react-router-ssr-query": "1.143.11",
-		"@tanstack/react-start": "1.143.11",
-		"@tanstack/react-table": "8.21.3",
-		"@tanstack/react-virtual": "3.13.13",
-		"@tanstack/table-core": "8.21.3",
-		"@tanstack/zod-adapter": "1.143.11",
-		"@types/node": "25.0.3",
-		"arctic": "3.7.0",
-		"consola": "3.4.2",
-		"discord-interactions": "4.4.0",
-		"drizzle-orm": "0.45.1",
-		"drizzle-zod": "0.8.3",
-		"googleapis": "169.0.0",
-		"graphql": "16.12.0",
-		"graphql-request": "7.4.0",
-		"hono": "4.11.2",
-		"keyv": "5.5.5",
-		"lucia": "3.2.2",
-		"luxon": "3.7.2",
-		"nanoid": "5.1.6",
-		"nanoid-dictionary": "5.0.0",
-		"path-to-regexp": "8.3.0",
-		"postgres": "3.4.7",
-		"react": "19.2.3",
-		"react-dom": "19.2.3",
-		"react-error-boundary": "6.0.0",
-		"react-hook-form": "7.69.0",
-		"react-icons": "5.5.0",
-		"ulidx": "2.4.1",
-		"usehooks-ts": "3.1.1",
-		"uuid": "13.0.0",
-		"zod": "4.2.1"
-	},
-	"devDependencies": {
-		"@biomejs/biome": "1.9.4",
-		"@changesets/changelog-github": "0.5.2",
-		"@changesets/cli": "2.29.8",
-		"@emotion/babel-plugin": "11.13.5",
-		"@emotion/cache": "11.14.0",
-		"@graphql-codegen/cli": "6.1.0",
-		"@graphql-codegen/typescript": "5.0.7",
-		"@graphql-codegen/typescript-graphql-request": "6.3.0",
-		"@graphql-codegen/typescript-operations": "5.0.7",
-		"@mui/lab": "7.0.1-beta.20",
-		"@types/bun": "1.3.5",
-		"@types/luxon": "3.7.1",
-		"@types/react": "19.2.7",
-		"@types/react-dom": "19.2.3",
-		"@vitejs/plugin-react": "5.1.2",
-		"babel-plugin-react-compiler": "19.1.0-rc.3",
-		"drizzle-kit": "0.31.8",
-		"knip": "5.77.1",
-		"lefthook": "2.0.12",
-		"typescript": "5.9.3",
-		"vite": "7.3.0",
-		"vite-tsconfig-paths": "6.0.3"
-	},
-	"trustedDependencies": [
-		"lefthook"
-	]
+  "name": "frontend",
+  "private": false,
+  "version": "5.14.16",
+  "type": "module",
+  "scripts": {
+    "dev": "bunx vite",
+    "build": "bun vite build",
+    "start": "PORT=1420 bun run .output/server/index.mjs",
+    "check": "bun biome check ./src",
+    "typecheck": "bun tsc --noEmit",
+    "knip": "knip-bun --config knip.json"
+  },
+  "main": ".output/server/index.mjs",
+  "files": [
+    "drizzle",
+    ".output"
+  ],
+  "dependencies": {
+    "@discordjs/builders": "1.13.1",
+    "@discordjs/core": "2.4.0",
+    "@discordjs/rest": "2.6.0",
+    "@emotion/react": "11.14.0",
+    "@emotion/styled": "11.14.1",
+    "@fontsource/roboto": "5.2.9",
+    "@graphql-yoga/subscription": "5.0.5",
+    "@hono/zod-validator": "0.7.6",
+    "@hookform/resolvers": "5.2.2",
+    "@keyv/postgres": "2.2.2",
+    "@lucia-auth/adapter-drizzle": "1.1.0",
+    "@mui/icons-material": "7.3.6",
+    "@mui/material": "7.3.6",
+    "@mui/x-date-pickers": "8.23.0",
+    "@t3-oss/env-core": "0.13.10",
+    "@tanstack/match-sorter-utils": "8.19.4",
+    "@tanstack/react-query": "5.90.12",
+    "@tanstack/react-query-devtools": "5.91.1",
+    "@tanstack/react-router": "1.143.11",
+    "@tanstack/react-router-devtools": "1.143.11",
+    "@tanstack/react-router-ssr-query": "1.143.11",
+    "@tanstack/react-start": "1.143.11",
+    "@tanstack/react-table": "8.21.3",
+    "@tanstack/react-virtual": "3.13.13",
+    "@tanstack/table-core": "8.21.3",
+    "@tanstack/zod-adapter": "1.143.11",
+    "@types/node": "25.0.3",
+    "arctic": "3.7.0",
+    "consola": "3.4.2",
+    "croner": "^9.1.0",
+    "discord-interactions": "4.4.0",
+    "drizzle-orm": "0.45.1",
+    "drizzle-zod": "0.8.3",
+    "googleapis": "169.0.0",
+    "hono": "4.11.2",
+    "keyv": "5.5.5",
+    "lucia": "3.2.2",
+    "luxon": "3.7.2",
+    "nanoid": "5.1.6",
+    "nanoid-dictionary": "5.0.0",
+    "path-to-regexp": "8.3.0",
+    "postgres": "3.4.7",
+    "react": "19.2.3",
+    "react-dom": "19.2.3",
+    "react-error-boundary": "6.0.0",
+    "react-hook-form": "7.69.0",
+    "react-icons": "5.5.0",
+    "ulidx": "2.4.1",
+    "usehooks-ts": "3.1.1",
+    "uuid": "13.0.0",
+    "zod": "4.2.1"
+  },
+  "devDependencies": {
+    "@biomejs/biome": "1.9.4",
+    "@changesets/changelog-github": "0.5.2",
+    "@changesets/cli": "2.29.8",
+    "@emotion/babel-plugin": "11.13.5",
+    "@emotion/cache": "11.14.0",
+    "@graphql-codegen/cli": "6.1.0",
+    "@graphql-codegen/typescript": "5.0.7",
+    "@graphql-codegen/typescript-graphql-request": "6.3.0",
+    "@graphql-codegen/typescript-operations": "5.0.7",
+    "@mui/lab": "7.0.1-beta.20",
+    "@types/bun": "1.3.5",
+    "@types/luxon": "3.7.1",
+    "@types/react": "19.2.7",
+    "@types/react-dom": "19.2.3",
+    "@vitejs/plugin-react": "5.1.2",
+    "babel-plugin-react-compiler": "19.1.0-rc.3",
+    "drizzle-kit": "0.31.8",
+    "knip": "5.77.1",
+    "lefthook": "2.0.12",
+    "typescript": "5.9.3",
+    "vite": "7.3.0",
+    "vite-tsconfig-paths": "6.0.3"
+  },
+  "trustedDependencies": [
+    "lefthook"
+  ]
 }
