--- src/gql/UpdateWebhook.gql
@@ -1,19 +1,0 @@
-mutation UpdateWebhook(
-	$id: ID!
-	$url: String
-	$cronExpression: String
-	$enabled: Boolean
-) {
-	updateCronWebhook(
-		id: $id
-		url: $url
-		cronExpression: $cronExpression
-		enabled: $enabled
-	) {
-		id
-		url
-		cronExpression
-		enabled
-		nextRun
-	}
-}
