--- src/gql/Webhook.gql
@@ -1,9 +1,0 @@
-query Webhook($id: ID!) {
-	cronWebhook(id: $id) {
-		id
-		url
-		cronExpression
-		enabled
-		nextRun
-	}
-}
