--- src/gql/Webhooks.gql
@@ -1,9 +1,0 @@
-query Webhooks {
-	cronWebhooks {
-		id
-		url
-		cronExpression
-		enabled
-		nextRun
-	}
-}
