+++ src/server.ts
@@ -4,14 +4,20 @@ import { type Register, createServerOnlyFn } from "@tanstack/react-start";
 import type { RequestOptions } from "@tanstack/react-start/server";
 import handler from "@tanstack/react-start/server-entry";
 import type { Server } from "bun";
+import { Cron, scheduledJobs } from "croner";
 import type { PostgresJsDatabase } from "drizzle-orm/postgres-js";
 import type { BunWebSocketData } from "hono/bun";
 import { Lucia, type RegisteredLucia } from "lucia";
+import { DateTime } from "luxon";
 import type z from "zod";
 import { getDB, migrate } from "./server/drizzle/db";
 import * as schema from "./server/drizzle/schema";
 import type { RSVPStatusSchema } from "./server/schema";
+import { reminderFn } from "./server/services/adminService";
 
+/**
+ * A pubsub events manager for future websocket features
+ */
 const pubSub = createPubSub<{
   "event:rsvpUpdated": [
     eventId: string,
@@ -23,25 +29,41 @@ const pubSub = createPubSub<{
   ];
 }>();
 
-type MyRequestContext = {
-  server: Server<BunWebSocketData>;
-  pubSub: typeof pubSub;
-  db: PostgresJsDatabase<typeof schema>;
-  lucia: RegisteredLucia;
-};
-
-// This needs to be tanstack router, currently incorrect
-declare module "@tanstack/react-start" {
-  interface Register {
-    server: {
-      requestContext: MyRequestContext;
-    };
-  }
-}
-
 await migrate();
 
 const db = await getDB();
+
+/**
+ * Restore CRON Jobs
+ */
+async function restoreCron() {
+  const filters = await db.query.eventParsingRuleTable.findMany();
+
+  for (const filter of filters) {
+    const existingJobs = scheduledJobs.filter((job) => job.name === filter.id);
+
+    for (const job of existingJobs) {
+      console.log(`Job: ${job.name} deleted`);
+      job.stop();
+    }
+
+    const job = new Cron(
+      filter.cronExpr,
+      {
+        name: filter.id,
+      },
+      reminderFn,
+    );
+
+    const nextRun = job.nextRun();
+
+    console.log(
+      `Job: ${job.name} (${filter.name}) created, next run: ${nextRun ? DateTime.fromJSDate(nextRun).toLocaleString(DateTime.DATETIME_MED) : "unknown"}`,
+    );
+  }
+}
+
+restoreCron();
 
 const adapter = new DrizzlePostgreSQLAdapter(
   db,
@@ -73,6 +95,22 @@ declare module "lucia" {
   interface Register {
     Lucia: typeof lucia;
     DatabaseUserAttributes: Omit<DatabaseUser, "id">;
+  }
+}
+
+type MyRequestContext = {
+  server: Server<BunWebSocketData>;
+  pubSub: typeof pubSub;
+  db: PostgresJsDatabase<typeof schema>;
+  lucia: RegisteredLucia;
+};
+
+// This needs to be tanstack router, currently incorrect
+declare module "@tanstack/react-start" {
+  interface Register {
+    server: {
+      requestContext: MyRequestContext;
+    };
   }
 }
